refactor(moderation): update pages import
