+++ ui/apps/moderation/src/pages/moderation-value.tsx
@@ -3,12 +3,12 @@ import { useParams } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
 
 import { RootComponentProps } from '@akashaorg/typings/ui';
-import ModerationValueCard from '@akashaorg/design-system-core/lib/components/ModerationValuesCard/value-card/index';
+import ModerationValueCard from '@akashaorg/design-system-core/lib/components/ModerationValuesCard/value-card';
 
 import { values } from '../services/values';
 import { externalLinks } from '../utils/external-links';
 
-const ValuePage: React.FC<RootComponentProps> = () => {
+export const ModerationValue: React.FC<RootComponentProps> = () => {
   const { t } = useTranslation('app-moderation-ewa');
 
   const { value } = useParams();
@@ -25,5 +25,3 @@ const ValuePage: React.FC<RootComponentProps> = () => {
     />
   );
 };
-
-export default ValuePage;
