+++ ui/apps/moderation/src/pages/moderators.tsx
@@ -11,7 +11,7 @@ import Spinner from '@akashaorg/design-system-core/lib/components/Spinner';
 
 const tabs = ['All', 'Active', 'Resigned', 'Revoked'];
 
-const AllModerators: React.FC<RootComponentProps> = () => {
+export const Moderators: React.FC<RootComponentProps> = () => {
   const [activeTab, setActiveTab] = React.useState<string>('All');
 
   const { t } = useTranslation('app-moderation-ewa');
@@ -71,5 +71,3 @@ const AllModerators: React.FC<RootComponentProps> = () => {
     </BasicCardBox>
   );
 };
-
-export default AllModerators;
